Add Show error message test

+++ src/hooks/useApplicationData.js
@@ -37,14 +37,19 @@ export default function useApplicaiton() {
   const dailyInterviewers = getInterviewersForDay(state, state.day);
 
   // Book, delete or edit interview
-  function bookInterview(id, interview, changeInSpots) {
+  function bookInterview(id, interview, changeInSpots, updateClient = false) {
     // Update the appointment and number of appointments being changed
     const appointment = {
       ...state.appointments[id],
       interview
     };
 
-    const days = [...state.days,];
+    if (updateClient) {
+
+      console.log(state.appointments);
+    }
+
+    const days = [...state.days];
     days[state.day - 1].spots += changeInSpots;
 
     // Update all appointments object
@@ -52,6 +57,11 @@ export default function useApplicaiton() {
       ...state.appointments,
       [id]: appointment
     };
+
+    if (updateClient) {
+      console.log(appointments.id)
+      return dispatch({ type: SET_INTERVIEW, appointments, days });
+    }
 
     // Update database
     const updateOrDelete = (interview) ? axios.put(`/api/appointments/${id}`, appointment) : axios.delete(`/api/appointments/${id}`)
@@ -67,18 +77,16 @@ export default function useApplicaiton() {
 
   // Get request to fetch all days
   useEffect(() => {
-    // const websocket = new WebSocket('ws://localhost:8000');
-    
-    // // Set up rules for the websocket
-    // websocket.onopen = () => websocket.send('ping'); // when the connection is established
-    // websocket.onmessage = (event) => console.log(event.data); // argument as the message, the function itself is what to do
-
     Promise.all([
       axios.get('/api/days'),
       axios.get('/api/appointments'),
       axios.get('/api/interviewers')
     ]).then((all) => {
       setApplicationData(all[0].data, all[1].data, all[2].data);
+      // const websocket = new WebSocket('ws://localhost:8001');
+      // websocket.onopen = () => {
+      //   websocket.onmessage = (event) => bookInterview(event.data.id, event.data.interview, (event.data.interview) ? 1 : -1, true);
+      // };
     })
   }, []);
